Unit Tests directory
This adds the unittests directory and some argus_timer tests to
kick things off.  unittests/unittest.h has some convenience macros
and functions which can be used to generate a uniform format for
reporting test results.

The unittests directory is NOT built by default.  To build:

% ./configure  # as normal
% ./config.status --file unittests/Makefile
% make
% make -C unittests

To build the unittests without building the rest of the distribution,
the "make" steps should instead be:

% make -C common
% make -C unittests