Add continuation paramater to ArgusParseTime()
This replaces the comparison of the "tm" parameter to the global
ArgusParser->RaLastFilter .  When "continued" is != 0, we are parsing
the second part of a time range string and the two struct tm
parameters should be swapped.  For example, when continued != 0
ArgusParseTime() expects to parse the five minutes in a string like
"-30m+5m".