radhcp: ArgusHandleSearchCommand searches interval tree
Clients send a string of the form "SEARCH: <timerange>", where
timerange is as described in the ra(1) man page for the -t option.

Parsing of the command string is a little weak and there must be
only one space between the colon and the timerange.

Allocate invecstr from heap and don't tell RabootpPrintDhcp() about the
full length to allow for ArgusPrintTime() running over the buffer's end.