exit immediately if any step of recursive make fails
The recursive make is a bourne shell script.  Use "set -e" to cause
the shell exit with a non-zero return value if any child process
does so.  Without this, if make fails in a subdirectory the parent
make process (a) continues to build other subdirectories and (b)
returns a value of 0 (zero) to the shell even though the build was
not successful.