Fix for RaParseCIDRAddr() to deal with 0.0.0.0/0 special case
When an address is passed to RaParseCIDRAddr() with a masklen of zero (default), we normally change that to length 32.  But if we pass 0.0.0.0/0, we would like to retain the zero masklen, as this can be used to select the root of the tree, rather than the terminal leaf address of 0.0.0.0/32.