add ArgusRealloc() and resize wire format buffers
When calling ArgusGenerateRecord() it is necessary to provide a buffer
sufficient in size to hold the largest possible flow record.  In practice,
flow records are almost never full size.  Reallocate the buffer after
the record has been formatted so that far less memory is used.

The Linux allocator is "optimistic" and, as a result, malloc() never
fails.  It appears that the wrong amount of allocated memory is reported
by top(1) (and others?) after oversubscribed allocations are freed.
It is not clear if this will cause confusion so rather than risk the rath
of the OOM killer reallocate buffers down to the size actually needed.

Oversubscribed allocations can be slower as a result of a mutex aquisition
in malloc.  One more reason to avoid them.

Trac ticket #76