build workstation RPMs with GeoIP support
GeoIP support was originally left out of the workstation builds so
that (a) the Linux workstation packages could have the same external
requirements as the OSX packages while (b) the OSX packages did not rely
on GeoIP which is not part of Apple's distribution.

To regain parity between the two, some work will be required to include
GeoIP libraries with the OSX build.

Update dependencies for -devel package to include GeoIP-devel and
some others.

GeoIP support requested by John Gerth.