radhcp: search results can be written to SQL
If a query is supplied on the radhcp commandline AND a mysql:// write
target is also supplied send search results to the database.

It is necessary to provide a time mode ("-M time xxx") when writing to
the database since this is used when generating table names if there is
a date component.