radhcp: simplify and correct inserts into interval trees
Do not ignore the return value of RB_INSERT called from
ArgusDhcpIntvlTreeInsert.  RB_INSERT searches the tree for the right
place to add the new node and returns a pointer to an existing node if
the key already exists in the tree.  This means that it is unnecessary
to search the tree first; instead, (1) up the cached record's refcount
(2) try to insert (3) decrement the refcount if insertion failed.