allow each query to save results in separate database table
Add the token "table" to the query string.  For example:
  -Mquery:SEARCH\ when=-1d,table=results_%Y_%m_%d

The date component will be filled in based on the end time of the "when="
range.  A write output paramter ("-w") specifying the MySQL user, host
and database name must be present when the db option is used in a query.

Example run with two queries, each written to its own table:

radhcp -r argus.out \
   -Mquery:SEARCH:\ when=-1d,table=detail_%Y_%m_%d \
   -Mquery:SEARCH:\ when=-1d,table=summary_%Y_%m_%d,pullup \
   -w mysql://root@localhost/dhcp/ \
   -M time 1d