radhcp: improve timezone lookup in nightly script
systemd provides the timedatectl utility to get/set clock information.
Use that to query the timezone and fall back to following the
/etc/localtime link if that doesn't work.

Remove the gnu-specific "-f" option passed to readlink so that this
script works outside of Linux.  An absolute path isn't necessary since
the awk script will still pick out the correct path elements.