argus_output.c: respond to client commands quickly
Processing client commands from either a client argus data socket or from
a control socket is split between two threads: ArgusListenProcess() and
__ArgusOutputProcess().  The listening thread marks the client readable
when select() returns and includes that client's file descriptor in the
"readfds" fd_set.  It also removes the file descriptor from the readfds
fd_set.  In response, the output thread reads the command, responds and
then marks the client unreadable.  Now the listening thread will again
select() on that file descriptor.

Since the listener's select times out after one second, noticing the next
client command can take up to one second.  Add a pair of pipe(2) file
descriptors to each output (ArgusOutputStruct) that the output thread can
use to signal the listener when it is done processing a client command.
One fd of each pair will also be added to the listener's select so that
it wakes up in a timely fashion.