ArgusWriteOutSocket: complete partial xmit from correct buffer
If the socket structure's "rec" field has data, be sure to set the
wire-format buffer pointer to that location.  Otherwise, whatever stale
(and possibly freed) value is left in that pointer will be used.

This problem is most noticeable when one or more connections are
encrypted by SASL.  Some SASL clients will be unable to decode data
received from radium and some non-SASL clients will receive data that
does not resemble flow records.