radhcp: expire bound transactions in client tree
Transactions that are not in the BOUND state are automatically removed
from the client tree after some short time.  However, no state transition
in radhcp is generated by the lease timer expiring.  Leases that expire
are never removed from the client tree.

Remove an expired lease from the client tree at the same time that it
is removed from the interval tree.