radhcp: handle DHCPRELEASE messages
A DHCP client can give up an active lease by sending a DHCPRELEASE
message.  Adjust the end time in the interval tree node for this lease,
advance the client state to INIT and mark the client structure "released".

Unfortunately, the transaction ID in the release message does not have
to match the transaction ID in any of the DHCP messages that established
the lease being discarded.  Search the patricia tree for the leased IP
address.  Only update a lease if the client hardware address that owns
the lease matches the client hardware address in the DHCPRELEASE message.
This will help avoid updating the wrong lease in the event of overlapping
address spaces.  It also provides some very very small protection against
another host forcing the lease to be terminated early.