radhcp: remove node from client tree when holddown expires
The holddown timer is what deicdes how long the node should stay in the
client (transaction) tree.  The refcount on the transaction node was
decremented by the lease/holddown timer callback but the node was left
in the tree.  The interval-tree timer expiration callback was removing
it from the tree (much later in most cases) and was _also_ decrementing
the refcount.