radhcp: intervaltree index now includes data pointer
Using the flow start time to distinguish lease intervals in the interval
tree is insufficient and experiences collisions with real-world data.
Accidentally updating an interval for some other lease not only corrupts
the victim lease's data, but also results in the reference count for the
offender's ArgusDhcpStruct being too small.  After the offending lease
expires the lease and interval timer callbacks will eventually try to
free the ArgusDhcpStruct twice, resulting in a segmentation fault.

Since the dhcp transactions are distinguished by both XID and client
MAC address, use the pointer to the ArgusDhcpStruct for a particular
transaction as a distinguishing characteristic since it effectively covers
both of those things without additional dereferencing and comparisons.