control channels do not output periodic MARs
The control channel is used to implement a protocol that is entirely
7-bit ascii text.  Periodically transmitting a binary management record
into this stream makes any automated use of the control channel very
difficult and adds little value.  Interpret a MAR update interval of
zero to mean "no updates" and set the interval to zero when creating
new control outputs.