radhcp: consider renewing and rebinding to be bound
For the purposes of timing out entries in the client tree, treat leases
that are currently being renewed as still bound.  No need to start
the discovery timer for a lease that left the BOUND state and is going
right back.

This can also prevent a client from being removed from the client tree
prematurely when we process packets out of order.  In some cases it is
impossible to tell from an aggregated flow record which of the "source"
and "dest" buffers was seen first on the wire.

Also make sure the timer in the dhcp structure is the timer for which
the __discover_exp_cb callback was made.