radhcp: disable discovery timer if lease ever bound
The reference aquired for the ArgusDhcpStruct at creation time
should be held until the structure is no longer needed anywhere.
This prevents the refcount from ever hitting zero before we're done
with it (i.e., somthing decrements the count and subsequently bumps
it back up).

If the lease has _ever_ been in the bound state we must hold this
reference until the interval tree(s) and patricia tree are done
with it.