print xml header only if record received from network
The XML header includes information about the data source that
generated the flow record.  If the record was generated internally
(not received from some other process) then the record does not
have a valid struct ArgusInput pointer.  Much of the xml header
will be meaningless; skip printing the header and only print the
record contents.