new ramanage utility
Add a utility for managing argus archive directories.  The three functions
provided by ramanage are compression, upload and removal.  All functions
can decide to act on a file based on its age.

This can add a new dependency on libcurl.  If libcurl is not found during
compile then ramange will attempt to run the curl executable for uploads,
instead.