workstation packages should not include analytics scripts
The nightly jobs to catalog hosts and ports and to look for scanners
should not be part of the workstation builds.  This is functionality
reserved for the collector and/or analytics appliances.  It also
introduces a dependency on the qosient::util perl module which is not
included in the base argus-clients-workstation package.

Trac ticket #122