always remove deleted client from output->ArgusClients queue
In the case that a client is disconnected before sending a START:
message, that client's ArgusClientData structure is never removed from
the clients queue.  If this happens enough the list fills to capacity
(limited by ARGUS_MAXCLIENTS) and no new clients are accepted, even
though none are actually connected at the time.

Add a "delete" flag to ArgusClientData, set it from ArgusDeleteSocket(),
and use this to determine when to remove a client from the list.