factor out directory creation into new function ArgusMkdirPath
This code is copied several places.  Consolidate into one location.
Allocate a copy of the path string on the heap instead of on the
stack.

Also clean up some of the surrounding code where ArgusMkdirPath is
now called.  Replace filename buffer allocations on stack with
memory allocated from heap.  Fix the usage of realpath() in some
places, which also helps to get rid of stack allocations.