add function to get SQL column names and primary key
RasqlManageGetColumns() retreives an array of column names and sorts
the array so that the columns that make up the primary key are first.
The output parameter keylen indicates how many of the array elements
make up the key.  The return value is the total number of column names
in the array.