rahisto: output histogram text and write flow records to file
Print the textual histogram data to stdout even if writing binary flow
records to a file, so long as the flow records are not also written to
stdout.  This will allow both representations of the data to be generated
with a single call to rahisto.  A shell script may take advantage of
this to write both columnar data and flow records into a database table.