build behavioral baseline data from rahisto
Introduce a set of programs to build, query and delete SQL tables
containing frequency distributions as generated by rahisto.

rahisto-updatesql will take the input filename from the FILENAME
environment variable if none is provided on the commandline, so it can
be run directly from rastream to process files as they are closed.

rahisto-querysql retrieves one or more histograms from the database.
When multiple histograms match the query parameters, they are aggregated
into a single histogram.  This program can also calculate some measure
of difference between to (possibly aggregated) histograms using the
Kullback-Leibler divergence.  This can be done for two specified time
ranges or on a pair-wise basis where every two consecutive histograms
within a time range are compared and the divergences are output as a
time series.

rahisto-delsql removes histograms from the database.  The time parameter
(-t) for this program indicates the _newest_ histograms to remain in the
database if no range is provided.  So "-t 2018/08/01" will remove all
histograms for argus records with timestamps before midnight 2018/08/01
UTC.