do not check for the Hstr option outside of rahisto
The ArgusParseArgs() function will only set parser->Hstr if the
client is rahisto.  No other clients need this so remove calls to
ArgusHistoMetricParse().

Some clients, such as ratop, attempt to use the presence of a non-NULL
string in Hstr to detect if the unary -H option was supplied.  Replace
these with tests for Hflag, instead.  This fixes ratop's ability to report
-H option at the top of the screen with the other commandline parameters.