introduce qosient::XS::util perl module
This adds a new compiled perl module to the base clients RPM.  The util
module has the ArgusParseTime() function which will eventually replace
qosient::util::parsetime().  Other functions can be added as needed.

In the common source directory, the ArgusParseTime() C function has
been moved into a separate file.  This file is compiled twice: once for
argus_common.a and again, with the -fPIC flag, for the perl module.

Perl scripts that need to parse date strings can use the new module with
something like:

  use qosient::XS::util;
  my ($result, $start, $stop) = qosient::XS::util('-1y+30d');