perllib: add depend target to Makefile.in
This target is always required when the .devel file is present.

Create dependencies for static source as well as the C source generated
by swig.  Wrap $^ with a filter expression in the Makefile to avoid
compiling all of the header file dependencies added by the 'depend'
target; this will fail since typically we don't compile as root and
can't write things into /usr/include.  The combination of $(filter)
and $^ allows VPATH to work as expected; do this instead of using $<.

Fix the include path in INCLS.

Fixes: 9b3bd4241568 ("introduce qosient::XS::util perl module")