rastream: do not flush files from child process
If the -f option is provided, but the program it specifies cannot
be executed (for any reason) execv() returns and the child rastream
is not replaced.  When the child exits, the stdio (FILE *)s are
flushed and closed.  Since these are shared with the parent process,
which is still writing to them, this will frequently cause the same
data to be written a file in multiple locations.  The resulting
file is unusable beyond that point.

Do not call exit(), either explicitly or implicitly, as this is the
function that flushes open files.  Instead, unceremoniously kill
off the child process if execv() fails.

Trac ticket #144