handle DOS-style newlines in RaParseResourceFile
Configuration files with unix-style newlines are difficult to deal
with on Windows systems as the available editors treat the entire
file as one line of text with invisible ("\n") characters sprinkled
throughout.  Update the main configuration parsing function to
recognize the "\r\n" end-of-line sequence so that configuration
files on windows systems can use the native text-file conventions.