ramanage: read configuration values from Windows registry
After reading configuration files fetch configuration values, if present,
from the Windows registry.  These take precedence over config items in
files.

All values are held in the registry key
HKLM:\SOFTWARE\QoSient\ArgusClients\ramanage

Any value that can be set in the configuration files can be the name of
a value in the ramanage key.

The -X option will cause configuration values not to be read from the
registry.