ramanage: limit how often the archive is processed
ramanage is intended to be run, primarily, from rastream each time it
closes a file.  Since rastream may have a number of files open and, in
the case of dividing up files by time, may close a number of files in
close succession, ramanage may be run several times over a very short
(seconds) interval.  If ramanage is configured to compress and upload,
this could result in significant cpu and/or network usage.

Add a configuration parameter that limits how often retained argus record
files are processed.  The minimum useful value will be one close to the
duration represented by the data in a single argus file.