rabins: process bins in a timely fashion when reading records from file
When reading from a file, the "clock" should be generated from flow
record timestamps, not from the wall clock.  The current time of day
should always be _after_ the time in a flow record being read from disk.
As a result, the entire file is read before any records are output.
If the file is sparse compared to the bin size, we may end up with one bin
per flow record, which results in one hash table (big) per flow record.