Fix error in ArgusFetchLastuSecTime.
ArgusFetchLastuSecTime used the time dsr subtype to decide what timestamps
to use to return the last time in the record.  Need to rely on the time dsr
to be correct and use both end times for the comparison.

ArgusGenerateRecordStruct needed to reset the time dsr subtype to indicate
that all fields are available for comparison.

Need to reorganize tests in ArgusGenerateRecordStruct to be more efficient.