common: do not use argus record input pointer after free
struct ArgusRecordStruct contains a pointer to the ArgusInput structure
that was the source of the flow record.  When reading files, this pointer
is no longer valid (has been freed) before RaParseComplete() is called.

The input pointer was used by ArgusInitAddrtoname() to get the local
subnet information.  Create a new function, ArgusSetLocalNet(), that
can be called as the files are processed.  Current behavior is preserved
such that the last input determines the local subnet.

Fixes: 5247c32301e4 ("common: defer initializing global name arrays")