radhcp: calc lease and interval tree expiries from time DSR
When the radhcp daemon is restarted, all previously accumulated lease
information is lost.  Calculate the lease and interval tree expirations
with the times found in the flow records containing the dhcp payloads
instead of setting the expiration as an offset from the current time of
day (i.e., now + lease time in seconds).  Now radhcp can first process
old flow records from disk or SQL, and then proceed to processing records
from a live stream.  This will "prime" the lease database so that queries
against existing leases can be made and so that new flow records can
update the existing leases.

For example,
  % radhcp -X -F radhcp.conf -r /tmp/cached -S ::1 -M control:10000

or
  % rasql -r mysql://root@localhost/dhcpFlows/dhcp_%Y_%m_%d -Mtime 1d \
          -t -24h -w - | radhcp -X -F radhcp.conf -r - -S ::1 -M control:10000

Trac ticket #125