fix buffer overrun in ArgusDebug
The calculated string length should be bounded by the buffer size.
The *snprintf() functions will return the number of bytes required to
format the string regardless of buffer size.  Check the return value of
vsnprintf() for the case where the debug string to log is larger than
the output buffer.

Ironically, this
Fixes: 9d56c6f39fb3 ("improve efficiency of ArgusPrintTime() and avoid buffer overruns")