rasql: exit with zero value if no tables match query
When the user specifies a table name (possibly with a strftime pattern
and time range) but no matching tables are found, output no data and
exit with success value.

In particular, this helps with starting radhcp on a fresh system install
or on a system that has been off for a day or more.  Loading the previous
24 hours on such a system will produce no data.

i.e.:
  rasql -r mysql://root@localhost/dhcpFlows/dhcp_%Y_%m_%d -Mtime 1d \
        -t -24h